LayoutTests/platform/wpe/media/video-muted-holds-sleep-assertion-expected.txt


RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (internals.elementIsBlockingDisplaySleep(video) == 'false'), OBSERVED 'true' FAIL
END OF TEST
